# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/webhooks/payment:
    post:
      operationId: WebhooksController_handlePaymentWebhook
      parameters:
        - name: x-webhook-signature
          required: true
          in: header
          schema:
            type: string
      responses:
        "201":
          description: ""
      summary: Webhook de paiement
      tags:
        - Webhooks
info:
  title: API École de Formation
  description: Gestion des étudiants, formateurs, cours et inscriptions
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...